+++ src/App.css
@@ -47,8 +47,9 @@ input:-webkit-autofill,
 input:-webkit-autofill:hover, 
 input:-webkit-autofill:focus, 
 input:-webkit-autofill:active {
-    -webkit-box-shadow: 0 0 0 30px #f2f2f2 inset;
+    -webkit-box-shadow: 0 0 0 1000px #f2f2f2 inset;
 }
+
 a {
   color: inherit;
   text-decoration: none;
