+++ src/components/FeedbackBox/FeedbackList.js
@@ -1,8 +1,6 @@
 /** @jsxImportSource @emotion/react */
 import { css } from '@emotion/react';
 import { useEffect, useState } from 'react';
-import { useRequestAuth } from '../../hooks/useRequestAuth';
-import { getApiEndpoint, isError, urlOwnerNotFound } from '../../utils/util';
 import FeedbackListCol from './FeedbackListCol';
 
 function shuffleArray(array) {
@@ -12,41 +10,16 @@ function shuffleArray(array) {
   }
 }
 
-function FeedbackList() {
-  const [feedbacks, setFeedbacks] = useState(null);
+function FeedbackList({ feedbacks }) {
   const [listDiv, setListdiv] = useState(<div />);
   const listDivOne = [];
   const listDivTwo = [];
   const listDivThree = [];
 
-  // 데이터 받아오는 코드 추가 예정
   useEffect(() => {
-    requestFeedbacks();
-  }, []);
-
-  const { res: feedbacksRes, request: requestFeedbacks } = useRequestAuth({
-    endpoint: `${getApiEndpoint()}/feedback`,
-    method: 'get',
-  });
-
-  useEffect(() => {
-    if (feedbacksRes && feedbacksRes.data) {
-      const { code, data, message } = feedbacksRes.data;
-      if (isError(code) && urlOwnerNotFound(message)) {
-        alert('정보를 가져올 수 없습니다.');
-      } else if (isError(code)) {
-        alert('데이터 베이스 에러입니다.');
-      }
-      if (data) {
-        setFeedbacks(data);
-      }
-    }
-  }, [feedbacksRes]);
-
-  useEffect(() => {
-    if (feedbacks !== null) {
+    if (feedbacks !== null && feedbacks !== undefined) {
       shuffleArray(feedbacks);
-      feedbacks.forEach((Feedback, index) => {
+      feedbacks.forEach((noData, index) => {
         if (index % 3 === 0) {
           listDivOne.push(feedbacks[index]);
         } else if (index % 3 === 1) {
@@ -55,19 +28,24 @@ function FeedbackList() {
           listDivThree.push(feedbacks[index]);
         }
       });
-      if (listDivOne.length !== 0) {
-        setListdiv(divideList());
-      }
     }
+    setListdiv(divideList());
   }, [feedbacks]);
 
   function divideList() {
+    let needFeedbackInfo = 0;
     return [listDivOne, listDivTwo, listDivThree].map((list, index) => {
       if (list.length) {
         return <FeedbackListCol key={list[0].feedback_seq} list={list} />;
-      }
-      const uniqKey = index * 1000;
-      return <FeedbackListCol key={uniqKey} list={[]} />;
+      } else needFeedbackInfo += 1;
+      const uniqKey = index * 10;
+      return (
+        <FeedbackListCol
+          key={uniqKey}
+          list={[]}
+          needFeedbackInfo={needFeedbackInfo}
+        />
+      );
     });
   }
 
@@ -80,7 +58,7 @@ const ContentBox = css`
   padding: 0 20px;
   display: flex;
   flex-wrap: wrap;
-  gap: 20px;
+  gap: 12.32px;
   box-sizing: border-box;
 `;
