+++ src/pages/JoinPage.js
@@ -1,17 +1,14 @@
 /** @jsxImportSource @emotion/react */
 import { css } from '@emotion/react';
-import {
-  COLOR_STYLE,
-  mq,
-  FlexSpaceBetweenCenter,
-} from '../styles/GlobalStyles';
+import { COLOR_STYLE, FlexSpaceBetweenCenter } from '../styles/GlobalStyles';
 
 import LeftBox from '../components/Join/LeftBox';
 import RightBox from '../components/Join/RightBox';
+import { MIN_WIDTH } from '../styles/style';
 
 function JoinPage() {
   return (
-    <div css={[Container, ContainerMQ()]}>
+    <div css={[Container]}>
       <LeftBox />
       <RightBox />
     </div>
@@ -22,16 +19,8 @@ export default JoinPage;
 
 const Container = css`
   width: 100%;
+  height: 100vh;
+  min-width: ${MIN_WIDTH};
   background-color: ${COLOR_STYLE.lightGrey};
   ${FlexSpaceBetweenCenter}
 `;
-
-export const ContainerMQ = () => {
-  const narrow = 'auto';
-  const normal = '100vh';
-
-  return mq({
-    flexDirection: ['column', 'column', 'row', 'row'],
-    height: [narrow, narrow, normal, normal],
-  });
-};
